requirements.txt
lxml>=2.3
cssselect
parsel
requests
requests-wsgi-adapter
nose
pyjade
werkzeug
requests-mock
httpbin
